requirements.txt
Flask
google-generativeai
redis
python-dotenv
requests